+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 4.0.0 (2019-12-30)
+
+- FIXED keep order of received dict and it's nested objects
+- Drops Python 2 support and Python 3.4
+
 ## 3.4.0 (2019-07-24)
 
 - FIXED sort order with regard to a nested list of dictionaries (#33 [wsantos](https://github.com/wsantos))